Fix formatting

+++ src/InertialSense.cpp
@@ -164,7 +164,7 @@ bool InertialSense::HasReceivedResponseFromDevice(size_t index)
 	}
 
 	printf("HasReceivedResponseFromDevice() %d %d %d %d\n",
-		index, 
+		(int)index, 
 		m_comManagerState.devices[index].flashCfg.size,
 		m_comManagerState.devices[index].devInfo.serialNumber,
 		m_comManagerState.devices[index].devInfo.manufacturer[0]);
